:truck: Moved utils files into utils folder in src
