+++ plugins/components/controllable/CMakeLists.txt
@@ -1,8 +1,8 @@
 # Add an executable for the server project
 add_library(Controllable SHARED
     Controllable.cpp
-    ../../systems/input/Keys.cpp
-    ../../systems/input/Actions.cpp
+    ../../../src/utils/Keys.cpp
+    ../../../src/utils/Actions.cpp
 )
 
 target_link_libraries(Controllable
